+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,5 @@
 <resources>
     <string name="app_name">tastezip</string>
-    <string name="client_id">1092102056221-92buirgtq5g5v53nbn3075r0167vgnan.apps.googleusercontent.com</string>
 
     <string name="naver_map_client_id">4ucf1ck4g1</string>
